Nullify stuff material effects on stats for gilded armor and headgear
Mousekin armor and headgear are mostly steel, with a thin gilded layer of material/"stuff".

As such, their hitpoints and work to make stats should not be affected by the material used for make them.

(Combat Extended also adds a mass multiplier for each material as well, so that needs to be nullified too)